Fix hash warnings.

+++ hash.c
@@ -154,7 +154,7 @@ uint32_t hash_str_jenkins(const void *key)
 /* Robert Jenkins's integer hash */
 uint32_t hash_int_jenkins(const void *key)
 {
-  register uint32_t a = (uint32_t)key;
+  register uint32_t a = (uintptr_t)key;
 
   a = (a+0x7ed55d16) + (a<<12);
   a = (a^0xc761c23c) ^ (a>>19);
@@ -169,7 +169,7 @@ uint32_t hash_int_jenkins(const void *key)
 /* Van Jacobson's IPv4 hash */
 uint32_t hash_int_jacobson(const void *key)
 {
-  register uint32_t hash = (uint32_t)key;
+  register uint32_t hash = (uintptr_t)key;
 
   hash ^= (hash >> 23) ^ (hash >> 17);
 
@@ -179,7 +179,7 @@ uint32_t hash_int_jacobson(const void *key)
 /* Knuth's multiplicative hash */
 uint32_t hash_int_knuth(const void *key)
 {
-  register uint32_t hash = (uint32_t)key;
+  register uint32_t hash = (uintptr_t)key;
 
   hash *= 0x9e3779b1;
